liteshop_bot/requirements.txt
aiogram>=3.7,<4.0
fastapi>=0.115,<1.0
python-dotenv>=1.0,<2.0
uvicorn[standard]>=0.30,<1.0
